backend/requirements.txt
asgiref                       3.7.2
certifi                       2023.5.7
cffi                          1.15.1
charset-normalizer            3.2.0
cryptography                  41.0.2
defusedxml                    0.7.1
diff-match-patch              20230430
Django                        3.2.20
django-filter                 23.2
django-import-export          3.2.0
django-templated-mail         1.1.1
djangorestframework           3.14.0
djangorestframework-simplejwt 5.2.2
djoser                        2.2.0
et-xmlfile                    1.1.0
flake8                        6.0.0
flake8-isort                  6.0.0
idna                          3.4
isort                         5.12.0
MarkupPy                      1.14
mccabe                        0.7.0
oauthlib                      3.2.2
odfpy                         1.4.1
openpyxl                      3.1.2
pip                           23.2.1
pycodestyle                   2.10.0
pycparser                     2.21
pyflakes                      3.0.1
PyJWT                         2.7.0
python-dotenv                 1.0.0
python3-openid                3.2.0
pytz                          2023.3
PyYAML                        6.0.1
requests                      2.31.0
requests-oauthlib             1.3.1
setuptools                    58.1.0
social-auth-app-django        5.2.0
social-auth-core              4.4.2
sqlparse                      0.4.4
tablib                        3.5.0
typing_extensions             4.7.1
urllib3                       2.0.3
xlrd                          2.0.1
xlwt                          1.3.0